refactor: re-export global types in index file
