+++ src/components/atoms/UiText/UiText.vue
@@ -9,8 +9,10 @@
 
 <script setup lang="ts">
 import type { HTMLAttributes } from 'vue';
-import type { DefineAttrs } from '../../../types/attrs';
-import type { HTMLTag } from '../../../types/tag';
+import type {
+  DefineAttrs,
+  HTMLTag,
+} from '../../../types';
 
 export interface TextProps {
   /**
