+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './attrs';
+export * from './icon';
+export * from './tag';
